refactor: rename MAC key ID to PRF key ID across API and internal services :hammer:
