+++ components/crm/api/swagger.yaml
@@ -402,8 +402,8 @@ definitions:
         description: The path to the Key Encryption Key in Vault.
         example: transit/keys/org-00000000-0000-0000-0000-000000000000
         type: string
-      mac_primary_key_id:
-        description: The primary key ID for MAC operations.
+      prf_primary_key_id:
+        description: The primary key ID for PRF-based search-token operations.
         example: 1
         type: integer
       organization_id:
