+++ components/crm/internal/services/encryption/audit.go
@@ -9,10 +9,12 @@ import (
 
 	libCommons "github.com/LerianStudio/lib-commons/v5/commons"
 	libLog "github.com/LerianStudio/lib-commons/v5/commons/log"
+	libHTTP "github.com/LerianStudio/lib-commons/v5/commons/net/http"
 	libOpenTelemetry "github.com/LerianStudio/lib-commons/v5/commons/opentelemetry"
 	libRuntime "github.com/LerianStudio/lib-commons/v5/commons/runtime"
 	"github.com/LerianStudio/midaz/v3/components/crm/internal/adapters/mongodb/audit"
 	"github.com/LerianStudio/midaz/v3/pkg/mmodel"
+	"go.opentelemetry.io/otel/attribute"
 )
 
 // AuditWriter emits protection audit events on a best-effort basis.
@@ -30,10 +32,32 @@ type AuditWriter interface {
 	EmitAsync(ctx context.Context, event *mmodel.ProtectionAuditEvent)
 }
 
+// AuditQueryService reads protection audit events for an organization.
+//
+// It is the read-side counterpart of AuditWriter: a thin, read-only
+// orchestration over the audit repository. The repository owns paging clamping,
+// sort defaulting, filtering, and opaque cursor handling; this service passes
+// the query straight through and returns the repository output unchanged so the
+// HTTP handler can map any error (including libHTTP.ErrInvalidCursor) to the
+// appropriate status code.
+type AuditQueryService interface {
+	// GetAuditEvents returns the audit events for organizationID matching query,
+	// together with the opaque next/prev cursor pagination produced by the
+	// repository. The query is forwarded verbatim — limit clamping and sort
+	// defaulting are the repository's responsibility. Any repository error is
+	// returned unchanged for the caller to translate to an HTTP status.
+	GetAuditEvents(ctx context.Context, organizationID string, query audit.AuditQuery) ([]*mmodel.ProtectionAuditEvent, libHTTP.CursorPagination, error)
+}
+
 // auditWriter is the repository-backed AuditWriter implementation.
 type auditWriter struct {
 	repo   audit.Repository
 	logger libLog.Logger
+}
+
+// auditQueryService is the repository-backed AuditQueryService implementation.
+type auditQueryService struct {
+	repo audit.Repository
 }
 
 // NewAuditWriter returns an AuditWriter backed by the given repository.
@@ -43,6 +67,13 @@ type auditWriter struct {
 // is no no-op fallback: a caller that has no repository has no AuditWriter.
 func NewAuditWriter(repo audit.Repository, logger libLog.Logger) AuditWriter {
 	return &auditWriter{repo: repo, logger: logger}
+}
+
+// NewAuditQueryService returns an AuditQueryService backed by the given audit
+// repository. The repository is supplied by the composition root and is always
+// a real, non-nil repository, so there is no nil guard.
+func NewAuditQueryService(repo audit.Repository) AuditQueryService {
+	return &auditQueryService{repo: repo}
 }
 
 // Emit writes the event synchronously. The repository error is intentionally
@@ -91,6 +122,38 @@ func (w *auditWriter) EmitAsync(ctx context.Context, event *mmodel.ProtectionAud
 		libRuntime.KeepRunning, func(c context.Context) {
 			w.Emit(c, event)
 		})
+}
+
+func (s *auditQueryService) GetAuditEvents(ctx context.Context, organizationID string, query audit.AuditQuery) ([]*mmodel.ProtectionAuditEvent, libHTTP.CursorPagination, error) {
+	logger, tracer, _, _ := libCommons.NewTrackingFromContext(ctx)
+
+	ctx, span := tracer.Start(ctx, "service.protection.get_audit_events")
+	defer span.End()
+
+	// Safe attributes only: org id and paging shape, plus which filters are set —
+	// never the filter VALUES (action/actor/outcome), which could carry context
+	// not appropriate for telemetry.
+	span.SetAttributes(
+		attribute.String("app.request.organization_id", organizationID),
+		attribute.Int("app.request.limit", query.Limit),
+		attribute.String("app.request.sort_order", query.SortOrder),
+		attribute.Bool("app.request.filter_action", query.Action != ""),
+		attribute.Bool("app.request.filter_actor", query.Actor != ""),
+		attribute.Bool("app.request.filter_outcome", query.Outcome != ""),
+	)
+
+	events, pagination, err := s.repo.FindByOrganization(ctx, organizationID, query)
+	if err != nil {
+		// Return the error unchanged: the handler maps repository errors
+		// (including libHTTP.ErrInvalidCursor) to the correct HTTP status.
+		libOpenTelemetry.HandleSpanError(span, "Failed to query audit events", err)
+
+		logger.Log(ctx, libLog.LevelWarn, "Failed to query audit events", libLog.Err(err))
+
+		return nil, libHTTP.CursorPagination{}, err
+	}
+
+	return events, pagination, nil
 }
 
 // safeAuditLogFields adapts the event's redacted field set to logger fields,
