+++ components/crm/internal/services/encryption/keyset_manager_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/LerianStudio/midaz/v3/pkg/crypto/kms/vault"
 	"github.com/LerianStudio/midaz/v3/pkg/crypto/tink"
 	"github.com/LerianStudio/midaz/v3/pkg/mmodel"
+	"github.com/LerianStudio/midaz/v3/tests/helpers"
 )
 
 // fakeKeysetRepo is a test double for mongoEncryption.KeysetRepository.
@@ -57,7 +58,7 @@ func (f *fakeKeysetRepo) getCalls() int {
 type fakeKeysetUnwrapper struct {
 	mu         sync.Mutex
 	aeadKeyset []byte
-	macKeyset  []byte
+	prfKeyset  []byte
 	err        error
 	calls      int
 	mountPaths []string // Records the mountPath received on each call.
@@ -74,13 +75,13 @@ func (f *fakeKeysetUnwrapper) UnwrapKeyset(_ context.Context, mountPath, _ strin
 		return nil, f.err
 	}
 
-	// Return AEAD keyset for AEAD wrapped data, MAC keyset for HMAC wrapped data
+	// Return AEAD keyset for AEAD wrapped data, PRF keyset for the compatibility slot.
 	if wrappedKeyset == "wrapped-aead" {
 		return f.aeadKeyset, nil
 	}
 
-	if wrappedKeyset == "wrapped-mac" {
-		return f.macKeyset, nil
+	if wrappedKeyset == "wrapped-prf" {
+		return f.prfKeyset, nil
 	}
 
 	return nil, errors.New("unknown wrapped keyset")
@@ -107,43 +108,35 @@ func (f *fakeKeysetUnwrapper) getMountPaths() []string {
 func generateTestKeysets(t *testing.T) ([]byte, []byte) {
 	t.Helper()
 
-	_, aeadBytes, err := tink.NewAEADKeysetGenerator().Generate()
-	if err != nil {
-		t.Fatalf("failed to generate AEAD keyset: %v", err)
-	}
+	keysets := helpers.GenerateTinkKeysets(t)
 
-	_, prfBytes, err := tink.NewPRFKeysetGenerator().Generate()
-	if err != nil {
-		t.Fatalf("failed to generate PRF keyset: %v", err)
-	}
-
-	return aeadBytes, prfBytes
+	return keysets.AEADBytes, keysets.PRFBytes
 }
 
 func TestKeysetManager_GetPrimitives_CacheMiss_Success(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
 			OrganizationID:    "org-123",
 			KEKPath:           "org-org-123",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
 
 	prims, err := manager.GetPrimitives(context.Background(), "org-123")
 	aead := prims.AEAD
-	mac := prims.PRF
+	prf := prims.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives() error = %v", err)
 	}
@@ -152,8 +145,8 @@ func TestKeysetManager_GetPrimitives_CacheMiss_Success(t *testing.T) {
 		t.Error("GetPrimitives() AEAD is nil")
 	}
 
-	if mac == nil {
-		t.Error("GetPrimitives() MAC is nil")
+	if prf == nil {
+		t.Error("GetPrimitives() PRF is nil")
 	}
 
 	if reader.getCalls() != 1 {
@@ -168,20 +161,20 @@ func TestKeysetManager_GetPrimitives_CacheMiss_Success(t *testing.T) {
 func TestKeysetManager_GetPrimitives_CacheHit_ReturnsCached(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
 			OrganizationID:    "org-456",
 			KEKPath:           "org-org-456",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -189,7 +182,7 @@ func TestKeysetManager_GetPrimitives_CacheHit_ReturnsCached(t *testing.T) {
 	// First call - cache miss
 	prims2, err := manager.GetPrimitives(context.Background(), "org-456")
 	aead1 := prims2.AEAD
-	mac1 := prims2.PRF
+	prf1 := prims2.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives() first call error = %v", err)
 	}
@@ -197,7 +190,7 @@ func TestKeysetManager_GetPrimitives_CacheHit_ReturnsCached(t *testing.T) {
 	// Second call - should be cache hit
 	prims3, err := manager.GetPrimitives(context.Background(), "org-456")
 	aead2 := prims3.AEAD
-	mac2 := prims3.PRF
+	prf2 := prims3.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives() second call error = %v", err)
 	}
@@ -207,8 +200,8 @@ func TestKeysetManager_GetPrimitives_CacheHit_ReturnsCached(t *testing.T) {
 		t.Error("GetPrimitives() AEAD primitives should be the same instance")
 	}
 
-	if mac1 != mac2 {
-		t.Error("GetPrimitives() MAC primitives should be the same instance")
+	if prf1 != prf2 {
+		t.Error("GetPrimitives() PRF primitives should be the same instance")
 	}
 
 	// Reader should only be called once
@@ -216,7 +209,7 @@ func TestKeysetManager_GetPrimitives_CacheHit_ReturnsCached(t *testing.T) {
 		t.Errorf("GetPrimitives() reader calls = %d, want 1", reader.getCalls())
 	}
 
-	// Unwrapper should only be called twice (once for AEAD, once for MAC on first call)
+	// Unwrapper should only be called twice (once for AEAD, once for PRF on first call)
 	if unwrapper.getCalls() != 2 {
 		t.Errorf("GetPrimitives() unwrapper calls = %d, want 2", unwrapper.getCalls())
 	}
@@ -225,20 +218,20 @@ func TestKeysetManager_GetPrimitives_CacheHit_ReturnsCached(t *testing.T) {
 func TestKeysetManager_GetPrimitives_CacheExpired_Refetches(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
 			OrganizationID:    "org-789",
 			KEKPath:           "org-org-789",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	// Use a very short TTL for testing
@@ -308,7 +301,7 @@ func TestKeysetManager_GetPrimitives_UnwrapError_Propagated(t *testing.T) {
 			OrganizationID:    "org-unwrap-fail",
 			KEKPath:           "org-org-unwrap-fail",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
@@ -337,14 +330,14 @@ func TestKeysetManager_GetPrimitives_ParseError_Propagated(t *testing.T) {
 			OrganizationID:    "org-parse-fail",
 			KEKPath:           "org-org-parse-fail",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	// Return invalid keyset bytes that cannot be parsed
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: []byte("invalid-keyset-data"),
-		macKeyset:  []byte("invalid-mac-keyset-data"),
+		prfKeyset:  []byte("invalid-prf-keyset-data"),
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -358,20 +351,20 @@ func TestKeysetManager_GetPrimitives_ParseError_Propagated(t *testing.T) {
 func TestKeysetManager_GetPrimitives_ConcurrentAccess_Safe(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
 			OrganizationID:    "org-concurrent",
 			KEKPath:           "org-org-concurrent",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -390,13 +383,13 @@ func TestKeysetManager_GetPrimitives_ConcurrentAccess_Safe(t *testing.T) {
 
 			prims4, err := manager.GetPrimitives(context.Background(), "org-concurrent")
 			aead := prims4.AEAD
-			mac := prims4.PRF
+			prf := prims4.PRF
 			if err != nil {
 				errChan <- err
 				return
 			}
 
-			if aead == nil || mac == nil {
+			if aead == nil || prf == nil {
 				errChan <- errors.New("nil primitives returned")
 			}
 		}()
@@ -418,7 +411,7 @@ func TestKeysetManager_GetPrimitives_ContextCancelled(t *testing.T) {
 			OrganizationID:    "org-ctx-cancel",
 			KEKPath:           "org-org-ctx-cancel",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
@@ -442,20 +435,20 @@ func TestKeysetManager_GetPrimitives_ContextCancelled(t *testing.T) {
 func TestKeysetManager_InvalidateCache_RemovesEntry(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
 			OrganizationID:    "org-invalidate",
 			KEKPath:           "org-org-invalidate",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -484,20 +477,20 @@ func TestKeysetManager_InvalidateCache_RemovesEntry(t *testing.T) {
 func TestKeysetManager_ClearCache_RemovesAllEntries(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
 			OrganizationID:    "org-clear",
 			KEKPath:           "org-org-clear",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -589,12 +582,12 @@ func TestDefaultKeysetManagerConfig(t *testing.T) {
 func TestKeysetManager_GetPrimitives_MultipleOrganizations(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{}
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -604,12 +597,12 @@ func TestKeysetManager_GetPrimitives_MultipleOrganizations(t *testing.T) {
 		OrganizationID:    "org-1",
 		KEKPath:           "org-org-1",
 		WrappedKeyset:     "wrapped-aead",
-		WrappedHMACKeyset: "wrapped-mac",
+		WrappedHMACKeyset: "wrapped-prf",
 	}
 
 	prims5, err := manager.GetPrimitives(context.Background(), "org-1")
 	aead1 := prims5.AEAD
-	mac1 := prims5.PRF
+	prf1 := prims5.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives(org-1) error = %v", err)
 	}
@@ -619,12 +612,12 @@ func TestKeysetManager_GetPrimitives_MultipleOrganizations(t *testing.T) {
 		OrganizationID:    "org-2",
 		KEKPath:           "org-org-2",
 		WrappedKeyset:     "wrapped-aead",
-		WrappedHMACKeyset: "wrapped-mac",
+		WrappedHMACKeyset: "wrapped-prf",
 	}
 
 	prims6, err := manager.GetPrimitives(context.Background(), "org-2")
 	aead2 := prims6.AEAD
-	mac2 := prims6.PRF
+	prf2 := prims6.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives(org-2) error = %v", err)
 	}
@@ -634,8 +627,8 @@ func TestKeysetManager_GetPrimitives_MultipleOrganizations(t *testing.T) {
 		t.Error("GetPrimitives() AEAD primitives should be different for different orgs")
 	}
 
-	if mac1 == mac2 {
-		t.Error("GetPrimitives() MAC primitives should be different for different orgs")
+	if prf1 == prf2 {
+		t.Error("GetPrimitives() PRF primitives should be different for different orgs")
 	}
 
 	// Reader should be called twice (once for each org)
@@ -690,7 +683,7 @@ func (f *slowKeysetRepo) getCalls() int {
 func TestKeysetManager_GetPrimitives_PerOrgMutex_DeduplicatesConcurrentFetches(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	// Use a slow reader to ensure concurrent requests overlap
 	reader := &slowKeysetRepo{
@@ -698,7 +691,7 @@ func TestKeysetManager_GetPrimitives_PerOrgMutex_DeduplicatesConcurrentFetches(t
 			OrganizationID:    "org-dedup",
 			KEKPath:           "org-org-dedup",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 		delay:      50 * time.Millisecond,
 		fetchStart: make(chan struct{}, 1),
@@ -706,7 +699,7 @@ func TestKeysetManager_GetPrimitives_PerOrgMutex_DeduplicatesConcurrentFetches(t
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -732,13 +725,13 @@ func TestKeysetManager_GetPrimitives_PerOrgMutex_DeduplicatesConcurrentFetches(t
 
 			prims7, err := manager.GetPrimitives(context.Background(), "org-dedup")
 			aead := prims7.AEAD
-			mac := prims7.PRF
+			prf := prims7.PRF
 			if err != nil {
 				errChan <- err
 				return
 			}
 
-			if aead == nil || mac == nil {
+			if aead == nil || prf == nil {
 				errChan <- errors.New("nil primitives returned")
 				return
 			}
@@ -766,7 +759,7 @@ func TestKeysetManager_GetPrimitives_PerOrgMutex_DeduplicatesConcurrentFetches(t
 		t.Errorf("GetPrimitives() reader calls = %d, want 1 (per-org mutex should deduplicate)", readerCalls)
 	}
 
-	// Unwrapper should only be called twice (once for AEAD, once for MAC)
+	// Unwrapper should only be called twice (once for AEAD, once for PRF)
 	unwrapperCalls := unwrapper.getCalls()
 	if unwrapperCalls != 2 {
 		t.Errorf("GetPrimitives() unwrapper calls = %d, want 2", unwrapperCalls)
@@ -924,7 +917,7 @@ func (f *fakeKeysetRepoWithProvision) markProvisioned() {
 func TestKeysetManager_GetPrimitives_AutoProvisionOnNotFound(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	// Reader returns not found initially, then returns keyset after provisioning
 	reader := &fakeKeysetRepoWithProvision{
@@ -934,14 +927,14 @@ func TestKeysetManager_GetPrimitives_AutoProvisionOnNotFound(t *testing.T) {
 			OrganizationID:    "org-auto-prov",
 			KEKPath:           "org-org-auto-prov",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 		errAfterProv: nil,
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	provisioner := &fakeProvisioningService{
@@ -949,7 +942,7 @@ func TestKeysetManager_GetPrimitives_AutoProvisionOnNotFound(t *testing.T) {
 			OrganizationID:   "org-auto-prov",
 			KEKPath:          "org-org-auto-prov",
 			AEADPrimaryKeyID: 12345,
-			MACPrimaryKeyID:  67890,
+			PRFPrimaryKeyID:  67890,
 		},
 		onProvision: func() {
 			reader.markProvisioned()
@@ -963,7 +956,7 @@ func TestKeysetManager_GetPrimitives_AutoProvisionOnNotFound(t *testing.T) {
 
 	prims8, err := manager.GetPrimitives(ctx, "org-auto-prov")
 	aead := prims8.AEAD
-	mac := prims8.PRF
+	prf := prims8.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives() error = %v", err)
 	}
@@ -972,8 +965,8 @@ func TestKeysetManager_GetPrimitives_AutoProvisionOnNotFound(t *testing.T) {
 		t.Error("GetPrimitives() AEAD is nil")
 	}
 
-	if mac == nil {
-		t.Error("GetPrimitives() MAC is nil")
+	if prf == nil {
+		t.Error("GetPrimitives() PRF is nil")
 	}
 
 	// Provisioner should be called once
@@ -990,20 +983,20 @@ func TestKeysetManager_GetPrimitives_AutoProvisionOnNotFound(t *testing.T) {
 func TestKeysetManager_GetPrimitives_NoAutoProvisionWhenKeysetExists(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
 			OrganizationID:    "org-existing",
 			KEKPath:           "org-org-existing",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	provisioner := &fakeProvisioningService{}
@@ -1012,7 +1005,7 @@ func TestKeysetManager_GetPrimitives_NoAutoProvisionWhenKeysetExists(t *testing.
 
 	prims9, err := manager.GetPrimitives(context.Background(), "org-existing")
 	aead := prims9.AEAD
-	mac := prims9.PRF
+	prf := prims9.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives() error = %v", err)
 	}
@@ -1021,8 +1014,8 @@ func TestKeysetManager_GetPrimitives_NoAutoProvisionWhenKeysetExists(t *testing.
 		t.Error("GetPrimitives() AEAD is nil")
 	}
 
-	if mac == nil {
-		t.Error("GetPrimitives() MAC is nil")
+	if prf == nil {
+		t.Error("GetPrimitives() PRF is nil")
 	}
 
 	// Provisioner should NOT be called when keyset exists
@@ -1125,7 +1118,7 @@ func TestKeysetManager_GetPrimitives_NilKeysetAfterProvision(t *testing.T) {
 func TestKeysetManager_autoProvision_UsesTenantFromContext(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	// Reader returns not found initially, then returns keyset after provisioning
 	reader := &fakeKeysetRepoWithProvision{
@@ -1135,14 +1128,14 @@ func TestKeysetManager_autoProvision_UsesTenantFromContext(t *testing.T) {
 			OrganizationID:    "org-tenant-ctx",
 			KEKPath:           "org-org-tenant-ctx",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 		errAfterProv: nil,
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	provisioner := &fakeProvisioningService{
@@ -1234,14 +1227,14 @@ func TestKeysetManager_autoProvision_DefaultsTenantWhenMissing(t *testing.T) {
 	t.Parallel()
 
 	// Setup keyset that will be returned after provisioning
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	provisionedKeyset := &mmodel.OrganizationKeyset{
 		TenantID:          "default",
 		OrganizationID:    "org-no-tenant",
 		KEKPath:           "transit/keys/crm/org-no-tenant",
 		WrappedKeyset:     "wrapped-aead",
-		WrappedHMACKeyset: "wrapped-mac",
+		WrappedHMACKeyset: "wrapped-prf",
 		KeysetInfo:        mmodel.KeysetInfo{PrimaryKeyID: 111},
 		HMACKeysetInfo:    mmodel.KeysetInfo{PrimaryKeyID: 222},
 	}
@@ -1254,7 +1247,7 @@ func TestKeysetManager_autoProvision_DefaultsTenantWhenMissing(t *testing.T) {
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	// Provisioner that simulates successful provisioning by updating the reader
@@ -1304,7 +1297,7 @@ func containsHelper(s, substr string) bool {
 func TestKeysetManager_GetPrimitives_TenantIsolation_CacheKeysScopedByTenant(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	// Track which organization IDs were requested
 	var readerMu sync.Mutex
@@ -1315,14 +1308,14 @@ func TestKeysetManager_GetPrimitives_TenantIsolation_CacheKeysScopedByTenant(t *
 			OrganizationID:    "same-org-id",
 			KEKPath:           "org-same-org-id",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	// Wrap the reader to track calls per context
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -1334,7 +1327,7 @@ func TestKeysetManager_GetPrimitives_TenantIsolation_CacheKeysScopedByTenant(t *
 	// First call from tenant A - should fetch from repo
 	prims10, err := manager.GetPrimitives(ctxTenantA, "same-org-id")
 	aeadA := prims10.AEAD
-	macA := prims10.PRF
+	prfA := prims10.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives(tenant-alpha) error = %v", err)
 	}
@@ -1346,7 +1339,7 @@ func TestKeysetManager_GetPrimitives_TenantIsolation_CacheKeysScopedByTenant(t *
 	// First call from tenant B - should fetch from repo (different cache key)
 	prims11, err := manager.GetPrimitives(ctxTenantB, "same-org-id")
 	aeadB := prims11.AEAD
-	macB := prims11.PRF
+	prfB := prims11.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives(tenant-beta) error = %v", err)
 	}
@@ -1359,7 +1352,7 @@ func TestKeysetManager_GetPrimitives_TenantIsolation_CacheKeysScopedByTenant(t *
 	// Second call from tenant A - should use cache
 	prims12, err := manager.GetPrimitives(ctxTenantA, "same-org-id")
 	aeadA2 := prims12.AEAD
-	macA2 := prims12.PRF
+	prfA2 := prims12.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives(tenant-alpha, second call) error = %v", err)
 	}
@@ -1370,7 +1363,7 @@ func TestKeysetManager_GetPrimitives_TenantIsolation_CacheKeysScopedByTenant(t *
 	}
 
 	// Tenant A should get same cached primitives
-	if aeadA != aeadA2 || macA != macA2 {
+	if aeadA != aeadA2 || prfA != prfA2 {
 		t.Error("GetPrimitives() tenant A should get same cached primitives on second call")
 	}
 
@@ -1380,8 +1373,8 @@ func TestKeysetManager_GetPrimitives_TenantIsolation_CacheKeysScopedByTenant(t *
 		t.Error("GetPrimitives() different tenants should have separate cache entries")
 	}
 
-	if macA == macB {
-		t.Error("GetPrimitives() different tenants should have separate MAC cache entries")
+	if prfA == prfB {
+		t.Error("GetPrimitives() different tenants should have separate PRF cache entries")
 	}
 }
 
@@ -1390,20 +1383,20 @@ func TestKeysetManager_GetPrimitives_TenantIsolation_CacheKeysScopedByTenant(t *
 func TestKeysetManager_InvalidateCacheForTenant_OnlyAffectsSpecificTenant(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
 			OrganizationID:    "shared-org",
 			KEKPath:           "org-shared-org",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -1475,7 +1468,7 @@ func TestKeysetManager_InvalidateCacheForTenant_OnlyAffectsSpecificTenant(t *tes
 func TestKeysetManager_GetPrimitives_AutoProvision_RegistryProperties(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	// Reader returns not found initially, then returns keyset after provisioning
 	reader := &fakeKeysetRepoWithProvision{
@@ -1485,14 +1478,14 @@ func TestKeysetManager_GetPrimitives_AutoProvision_RegistryProperties(t *testing
 			OrganizationID:    "org-registry-test",
 			KEKPath:           "org-org-registry-test",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 		errAfterProv: nil,
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	// Provisioner that returns result with expected registry properties
@@ -1501,7 +1494,7 @@ func TestKeysetManager_GetPrimitives_AutoProvision_RegistryProperties(t *testing
 			OrganizationID:   "org-registry-test",
 			KEKPath:          "org-org-registry-test",
 			AEADPrimaryKeyID: 12345,
-			MACPrimaryKeyID:  67890,
+			PRFPrimaryKeyID:  67890,
 			RegistryStatus:   mmodel.RegistryStatusActive,
 		},
 		onProvision: func() {
@@ -1516,7 +1509,7 @@ func TestKeysetManager_GetPrimitives_AutoProvision_RegistryProperties(t *testing
 
 	prims17, err := manager.GetPrimitives(ctx, "org-registry-test")
 	aead := prims17.AEAD
-	mac := prims17.PRF
+	prf := prims17.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives() error = %v", err)
 	}
@@ -1525,8 +1518,8 @@ func TestKeysetManager_GetPrimitives_AutoProvision_RegistryProperties(t *testing
 		t.Error("GetPrimitives() AEAD is nil")
 	}
 
-	if mac == nil {
-		t.Error("GetPrimitives() MAC is nil")
+	if prf == nil {
+		t.Error("GetPrimitives() PRF is nil")
 	}
 
 	// Verify provisioner was called
@@ -1655,21 +1648,21 @@ func TestCachedPrimitives_MultiKeyPRF(t *testing.T) {
 func TestKeysetManager_GetPrimitives_PopulatesMultiKeyPRF(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
 			OrganizationID:    "org-multikey",
 			KEKPath:           "org-org-multikey",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 			KeysetInfo:        mmodel.KeysetInfo{PrimaryKeyID: 111},
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -1677,7 +1670,7 @@ func TestKeysetManager_GetPrimitives_PopulatesMultiKeyPRF(t *testing.T) {
 	// Call GetPrimitives to trigger fetchAndCache
 	prims18, err := manager.GetPrimitives(context.Background(), "org-multikey")
 	aead := prims18.AEAD
-	mac := prims18.PRF
+	prf := prims18.PRF
 	if err != nil {
 		t.Fatalf("GetPrimitives() error = %v", err)
 	}
@@ -1685,8 +1678,8 @@ func TestKeysetManager_GetPrimitives_PopulatesMultiKeyPRF(t *testing.T) {
 	if aead == nil {
 		t.Error("GetPrimitives() AEAD is nil")
 	}
-	if mac == nil {
-		t.Error("GetPrimitives() MAC is nil")
+	if prf == nil {
+		t.Error("GetPrimitives() PRF is nil")
 	}
 
 	// Verify MultiKeyPRF is populated in the cache
@@ -1725,7 +1718,7 @@ func TestKeysetManager_fetchAndCache_MultiKeyPRF_StrictErrorMode(t *testing.T) {
 			OrganizationID:    "org-strict-error",
 			KEKPath:           "org-org-strict-error",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 			KeysetInfo:        mmodel.KeysetInfo{PrimaryKeyID: 111},
 		},
 	}
@@ -1738,7 +1731,7 @@ func TestKeysetManager_fetchAndCache_MultiKeyPRF_StrictErrorMode(t *testing.T) {
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  []byte("invalid-prf-keyset-data"),
+		prfKeyset:  []byte("invalid-prf-keyset-data"),
 	}
 
 	manager := NewKeysetManager(reader, unwrapper, nil, DefaultKeysetManagerConfig(), NewProtectionMetrics(nil))
@@ -1762,11 +1755,11 @@ func TestKeysetManager_fetchAndCache_MultiKeyPRF_StrictErrorMode(t *testing.T) {
 
 // TestKeysetManager_GetPrimitives_UnwrapMount_DefaultTenant_Flat verifies that a stored
 // keyset with TenantID "default" resolves to the flat base mount ("transit") for both the
-// AEAD and MAC unwrap calls. The mount is derived from the STORED keyset.TenantID, not ctx.
+// AEAD and PRF unwrap calls. The mount is derived from the STORED keyset.TenantID, not ctx.
 func TestKeysetManager_GetPrimitives_UnwrapMount_DefaultTenant_Flat(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	reader := &fakeKeysetRepo{
 		keyset: &mmodel.OrganizationKeyset{
@@ -1774,13 +1767,13 @@ func TestKeysetManager_GetPrimitives_UnwrapMount_DefaultTenant_Flat(t *testing.T
 			OrganizationID:    "org-flat",
 			KEKPath:           "crm/org-flat",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	config := KeysetManagerConfig{BaseMountPath: "transit"}
@@ -1805,11 +1798,11 @@ func TestKeysetManager_GetPrimitives_UnwrapMount_DefaultTenant_Flat(t *testing.T
 
 // TestKeysetManager_GetPrimitives_UnwrapMount_NonDefaultTenant_SubMount verifies that a
 // stored keyset with a non-default TenantID resolves to a per-tenant sub-mount
-// "transit/<tenant>" for both AEAD and MAC unwrap calls, derived from the STORED tenant.
+// "transit/<tenant>" for both AEAD and PRF unwrap calls, derived from the STORED tenant.
 func TestKeysetManager_GetPrimitives_UnwrapMount_NonDefaultTenant_SubMount(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	const tenant = "11111111-2222-3333-4444-555555555555"
 
@@ -1819,13 +1812,13 @@ func TestKeysetManager_GetPrimitives_UnwrapMount_NonDefaultTenant_SubMount(t *te
 			OrganizationID:    "org-sub",
 			KEKPath:           "crm/org-sub",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	config := KeysetManagerConfig{BaseMountPath: "transit"}
@@ -1865,7 +1858,7 @@ func TestKeysetManager_GetPrimitives_MountNotFound_FailsClosed(t *testing.T) {
 			OrganizationID:    "org-no-mount",
 			KEKPath:           "crm/org-no-mount",
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
@@ -1888,11 +1881,11 @@ func TestKeysetManager_GetPrimitives_MountNotFound_FailsClosed(t *testing.T) {
 
 // TestKeysetManager_GetPrimitives_StoredMountWins verifies that when the fetched
 // keyset carries a stored KEKMountPath, unwrap uses that mount for BOTH AEAD and
-// MAC, even when the manager's BaseMountPath would derive a different value.
+// PRF, even when the manager's BaseMountPath would derive a different value.
 func TestKeysetManager_GetPrimitives_StoredMountWins(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	const storedMount = "transit/tenant-x"
 
@@ -1903,13 +1896,13 @@ func TestKeysetManager_GetPrimitives_StoredMountWins(t *testing.T) {
 			KEKPath:           "crm/org-stored",
 			KEKMountPath:      storedMount,
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	// BaseMountPath deliberately differs from the stored mount's base.
@@ -1941,7 +1934,7 @@ func TestKeysetManager_GetPrimitives_StoredMountWins(t *testing.T) {
 func TestKeysetManager_GetPrimitives_ConfigBaseChanged_UsesStoredMount(t *testing.T) {
 	t.Parallel()
 
-	aeadBytes, macBytes := generateTestKeysets(t)
+	aeadBytes, prfBytes := generateTestKeysets(t)
 
 	const provisionedMount = "transit"
 
@@ -1952,13 +1945,13 @@ func TestKeysetManager_GetPrimitives_ConfigBaseChanged_UsesStoredMount(t *testin
 			KEKPath:           "crm/org-config-changed",
 			KEKMountPath:      provisionedMount,
 			WrappedKeyset:     "wrapped-aead",
-			WrappedHMACKeyset: "wrapped-mac",
+			WrappedHMACKeyset: "wrapped-prf",
 		},
 	}
 
 	unwrapper := &fakeKeysetUnwrapper{
 		aeadKeyset: aeadBytes,
-		macKeyset:  macBytes,
+		prfKeyset:  prfBytes,
 	}
 
 	// Live config base changed after provisioning.
@@ -2003,7 +1996,7 @@ func TestKeysetManager_GetPrimitives_LegacyEmptyMount_FallsBackToDerived(t *test
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			aeadBytes, macBytes := generateTestKeysets(t)
+			aeadBytes, prfBytes := generateTestKeysets(t)
 
 			reader := &fakeKeysetRepo{
 				keyset: &mmodel.OrganizationKeyset{
@@ -2012,13 +2005,13 @@ func TestKeysetManager_GetPrimitives_LegacyEmptyMount_FallsBackToDerived(t *test
 					KEKPath:           "crm/org-legacy",
 					KEKMountPath:      "", // legacy record predating the stored mount
 					WrappedKeyset:     "wrapped-aead",
-					WrappedHMACKeyset: "wrapped-mac",
+					WrappedHMACKeyset: "wrapped-prf",
 				},
 			}
 
 			unwrapper := &fakeKeysetUnwrapper{
 				aeadKeyset: aeadBytes,
-				macKeyset:  macBytes,
+				prfKeyset:  prfBytes,
 			}
 
 			config := KeysetManagerConfig{BaseMountPath: "transit"}
