+++ pkg/crypto/tink/legacy_import_test.go
@@ -5,7 +5,10 @@
 package tink
 
 import (
+	"crypto/hmac"
+	"crypto/sha256"
 	"encoding/base64"
+	"encoding/hex"
 	"testing"
 
 	libCrypto "github.com/LerianStudio/lib-commons/v5/commons/crypto"
@@ -59,17 +62,67 @@ func TestLegacyAESGCMPrimitiveFromHexKey_EncryptsLibCommonsCompatibleCiphertext(
 	assert.Equal(t, plaintext, *opened)
 }
 
-func TestLegacyMACPrimitiveFromSecret_ProducesLibCommonsTokenExactly(t *testing.T) {
+// TestLegacyPRFPrimitiveFromSecret_ByteEqualsLibCommonsAndLegacyHMAC is the HARD GATE
+// for T-2.1.4: the PRF-backed legacy primitive MUST produce, for the same secret
+// and value, a hex token byte-identical to BOTH (a) lib-commons GenerateHash and
+// (b) the prior HMAC-SHA256(secret, value) token construction. That legacy
+// construction is asserted against a reference HMAC computed inline with
+// crypto/hmac. Covers empty string and unicode.
+func TestLegacyPRFPrimitiveFromSecret_ByteEqualsLibCommonsAndLegacyHMAC(t *testing.T) {
 	t.Parallel()
 
 	legacy := &libCrypto.Crypto{HashSecretKey: testLegacyHashKey, EncryptSecretKey: testLegacyEncryptHexKey}
-	normalized := "abc123"
-	want := legacy.GenerateHash(&normalized)
 
-	primitive, err := NewLegacyMACPrimitiveFromSecret(testLegacyHashKey)
+	prfPrimitive, err := NewLegacyPRFPrimitiveFromSecret(testLegacyHashKey)
 	require.NoError(t, err)
 
-	got, err := primitive.ComputeLegacyHexToken([]byte(normalized))
-	require.NoError(t, err)
-	assert.Equal(t, want, got)
+	values := []string{
+		"abc123",
+		"",
+		"José da Silva — 日本語 🔐",
+		"a@b.com",
+	}
+
+	for _, value := range values {
+		value := value
+
+		t.Run("value="+value, func(t *testing.T) {
+			t.Parallel()
+
+			wantLibCommons := legacy.GenerateHash(&value)
+
+			// (b) reference: raw HMAC-SHA256(secret, value) hex, matching the
+			// removed legacy HMAC construction (RAW HmacKey, SHA256, TagSize 32).
+			refHMAC := hmac.New(sha256.New, []byte(testLegacyHashKey))
+			_, err := refHMAC.Write([]byte(value))
+			require.NoError(t, err)
+			wantHMAC := hex.EncodeToString(refHMAC.Sum(nil))
+
+			gotPRF, err := prfPrimitive.ComputeLegacyHexToken([]byte(value))
+			require.NoError(t, err)
+
+			// (a) PRF == lib-commons GenerateHash (the abort gate)
+			assert.Equal(t, wantLibCommons, gotPRF, "PRF token must byte-match lib-commons GenerateHash")
+			// (b) PRF == prior legacy HMAC construction
+			assert.Equal(t, wantHMAC, gotPRF, "PRF token must byte-match prior HMAC-SHA256 token output")
+			// 64-char lowercase hex (full HMAC-SHA256 tag)
+			assert.Len(t, gotPRF, 64, "legacy hex token must be 64 chars (32-byte HMAC-SHA256)")
+		})
+	}
+}
+
+func TestLegacyPRFPrimitiveFromSecret_EmptySecretRejected(t *testing.T) {
+	t.Parallel()
+
+	_, err := NewLegacyPRFPrimitiveFromSecret("")
+	require.Error(t, err)
+}
+
+func TestLegacyPRFPrimitive_NilGuard(t *testing.T) {
+	t.Parallel()
+
+	var primitive *LegacyPRFPrimitive
+
+	_, err := primitive.ComputeLegacyHexToken([]byte("x"))
+	require.Error(t, err)
 }
